refactor(tests): rename WebApi.Tests to WebApi.UnitTests
