+++ src/KRAFT.Results.WebApi/KRAFT.Results.WebApi.csproj
@@ -6,7 +6,7 @@
 
   <ItemGroup>
     <InternalsVisibleTo Include="KRAFT.Results.WebApi.IntegrationTests" />
-    <InternalsVisibleTo Include="KRAFT.Results.WebApi.Tests" />
+    <InternalsVisibleTo Include="KRAFT.Results.WebApi.UnitTests" />
   </ItemGroup>
 
   <ItemGroup>
