+++ tests/KRAFT.Results.WebApi.UnitTests/ValueObjects/EmailTests/EqualityOperators.cs
@@ -2,7 +2,7 @@ using KRAFT.Results.WebApi.ValueObjects;
 
 using Shouldly;
 
-namespace KRAFT.Results.WebApi.Tests.ValueObjects.EmailTests;
+namespace KRAFT.Results.WebApi.UnitTests.ValueObjects.EmailTests;
 
 public sealed class EqualityOperators
 {
