+++ tests/KRAFT.Results.WebApi.UnitTests/ValueObjects/EmailTests/GetHashCode.cs
@@ -2,15 +2,15 @@ using KRAFT.Results.WebApi.ValueObjects;
 
 using Shouldly;
 
-namespace KRAFT.Results.WebApi.Tests.ValueObjects.SlugTests;
+namespace KRAFT.Results.WebApi.UnitTests.ValueObjects.EmailTests;
 
 public sealed class GetHashCode
 {
     [Fact]
     public void ReturnsSameHashCode_WhenSameValue()
     {
-        Slug a = Slug.Create("hello-world");
-        Slug b = Slug.Create("hello-world");
+        Email a = Email.Create("test@example.com").FromResult();
+        Email b = Email.Create("test@example.com").FromResult();
 
         a.GetHashCode().ShouldBe(b.GetHashCode());
     }
