+++ tests/KRAFT.Results.WebApi.UnitTests/ValueObjects/PasswordTests/Equals.cs
@@ -2,7 +2,7 @@ using KRAFT.Results.WebApi.ValueObjects;
 
 using Shouldly;
 
-namespace KRAFT.Results.WebApi.Tests.ValueObjects.PasswordTests;
+namespace KRAFT.Results.WebApi.UnitTests.ValueObjects.PasswordTests;
 
 public sealed class Equals
 {
