+++ tests/KRAFT.Results.WebApi.UnitTests/ValueObjects/SlugTests/EqualityOperators.cs
@@ -2,15 +2,15 @@ using KRAFT.Results.WebApi.ValueObjects;
 
 using Shouldly;
 
-namespace KRAFT.Results.WebApi.Tests.ValueObjects.GenderTests;
+namespace KRAFT.Results.WebApi.UnitTests.ValueObjects.SlugTests;
 
 public sealed class EqualityOperators
 {
     [Fact]
     public void EqualOperator_ReturnsTrue_WhenSameValue()
     {
-        Gender a = Gender.Male;
-        Gender b = Gender.Parse("m");
+        Slug a = Slug.Create("hello-world");
+        Slug b = Slug.Create("hello-world");
 
         (a == b).ShouldBeTrue();
     }
@@ -18,6 +18,9 @@ public sealed class EqualityOperators
     [Fact]
     public void InequalityOperator_ReturnsTrue_WhenDifferentValue()
     {
-        (Gender.Male != Gender.Female).ShouldBeTrue();
+        Slug a = Slug.Create("hello");
+        Slug b = Slug.Create("world");
+
+        (a != b).ShouldBeTrue();
     }
 }
